lab6_streamlit_dashboard/pyproject.toml
[tool.poetry]
name = "lab6-streamlit-dashboard"
version = "0.1.0"
description = "Interactive ML Dashboard with Streamlit and MLflow"
authors = ["Olena Kirieva <Olena.Kirieva@cs.khpi.edu.ua>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
torch = {version = "^2.5.1", source = "pytorch"}
torchvision = {version = "^0.20.1", source = "pytorch"}
tqdm = "^4.66.0"
matplotlib = "^3.8.0"
numpy = "^1.26.0"
pyyaml = "^6.0.1"
pandas = "^2.2.0"
scikit-learn = "^1.4.0"
pillow = "^10.2.0"
requests = "^2.33.1"
mlflow = "^3.12.0"
streamlit = "^1.32.0"
plotly = "^5.18.0"
grad-cam = "^1.5.0"
lime = "^0.2.0.1"

[tool.poetry.group.dev.dependencies]
mypy = "^1.9.0"
ruff = "^0.3.0"
black = "^24.2.0"
isort = "^5.13.0"
types-pyyaml = "^6.0.12.20240311"
types-tqdm = "^4.66.0.20240106"
types-requests = "^2.31.0.20240311"
pandas-stubs = "^2.2.0.240303"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu121"
priority = "supplemental"